cleaner version/class
