+++ src/abstract_camera.py
